react_chat v0.0.1 ReactChat.Router.Helpers

Module with named helpers generated from ReactChat.Router.

Functions

api_auth_path(conn_or_endpoint, action)

api_auth_path(conn_or_endpoint, action, params)

api_auth_path(conn_or_endpoint, action, id, params)

api_auth_url(conn_or_endpoint, action)

api_auth_url(conn_or_endpoint, action, params)

api_auth_url(conn_or_endpoint, action, id, params)

api_registration_path(conn_or_endpoint, action)

api_registration_path(conn_or_endpoint, action, params)

api_registration_url(conn_or_endpoint, action)

api_registration_url(conn_or_endpoint, action, params)

api_user_path(conn_or_endpoint, action)

api_user_path(conn_or_endpoint, action, params)

api_user_path(conn_or_endpoint, action, id, params)

api_user_url(conn_or_endpoint, action)

api_user_url(conn_or_endpoint, action, params)

api_user_url(conn_or_endpoint, action, id, params)

page_path(conn_or_endpoint, action, path)

page_path(conn_or_endpoint, action, path, params)

page_url(conn_or_endpoint, action, path)

page_url(conn_or_endpoint, action, path, params)

path(data, path)

Generates the path information including any necessary prefix.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.