react_chat v0.0.1 ReactChat.User

Functions

changeset(struct, params \\ %{})

Builds a changeset based on the struct and params.

registration_changeset(struct, params \\ %{})

Is used for user creation. Validates presence of password and hashes it.

update_changeset(struct, params \\ %{})

Changeset that is used for user update and does not validate presence of password but if present hashes it.